[Refactor] seller 패키지 구조 리팩터링 진행 1. Seller/seller 구조로 변경 진행
